Improve Coord2d/3d ToString(), fix Coord2d/3d == and != (+ add tests), display mouse position in debug mode, improve samples

+++ Nzy3d/Plot3D/Primitives/Axes/Layout/Renderers/ITickRenderer.cs
@@ -2,13 +2,6 @@ namespace Nzy3d.Plot3D.Primitives.Axes.Layout.Renderers
 {
 	public interface ITickRenderer
 	{
-		string Format(float value);
+		string Format(double value);
 	}
 }
-
-//=======================================================
-//Service provided by Telerik (www.telerik.com)
-//Conversion powered by NRefactory.
-//Twitter: @telerik
-//Facebook: facebook.com/telerik
-//=======================================================
