+++ Nzy3d/Plot3D/Primitives/Axes/Layout/Renderers/IntegerTickRenderer.cs
@@ -2,16 +2,9 @@ namespace Nzy3d.Plot3D.Primitives.Axes.Layout.Renderers
 {
 	public class IntegerTickRenderer : ITickRenderer
 	{
-		public string Format(float value)
+		public string Format(double value)
 		{
 			return Convert.ToInt32(value).ToString();
 		}
 	}
 }
-
-//=======================================================
-//Service provided by Telerik (www.telerik.com)
-//Conversion powered by NRefactory.
-//Twitter: @telerik
-//Facebook: facebook.com/telerik
-//=======================================================
